[State Sync] Move executor proxy unit tests out of integration tests folder.
